final_project/airflow/requirements.txt
absl-py==0.12.0
alembic==1.7.7
anyio==3.5.0
apache-airflow==1.10.14
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==1.3.3
appnope==0.1.2
arch==4.19
argcomplete==1.12.3
argon2-cffi==20.1.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
black==22.1.0
bleach==3.3.0
blinker==1.4
bokeh==2.3.2
cached-property==1.5.2
cachelib==0.6.0
cachetools==5.0.0
cattrs==1.10.0
certifi==2021.10.8
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.12
cli-helpers==2.2.1
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.0.2
commonmark==0.9.1
configobj==5.0.6
configparser==3.5.3
connexion==2.13.0
croniter==0.3.37
cryptography==36.0.2
cvxpy==1.1.11
cycler==0.10.0
Cython==0.29.22
data==0.4
databricks-cli==0.16.2
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.4
dnspython==2.2.1
docutils==0.16
ecos==2.0.7.post1
email-validator==1.1.3
empyrical==0.5.5
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
filelock==3.4.2
Flask==1.1.4
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
Flask-WTF==0.14.3
flatbuffers==1.12
fsspec==2022.2.0
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
gcloud==0.18.3
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
google-cloud-bigquery==2.34.2
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.12.0
h5py==2.10.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
humanize==3.14.0
idna==2.10
importlib-metadata==4.11.3
importlib-resources==1.5.0
inflection==0.5.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipyturtle==0.2.4
ipywidgets==7.6.3
iso8601==1.0.2
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json-merge-patch==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
latex==0.7.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.3
Mako==1.2.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.4.0
matplotlib-inline==0.1.3
missingno==0.5.0
mistune==0.8.4
mobilechelonian==0.5
mpmath==1.2.1
multitasking==0.0.9
mypy-extensions==0.4.3
natsort==8.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.4.8
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.7
opt-einsum==3.3.0
optim==0.1.0
osqp==0.6.2.post0
packaging==20.9
pandas==1.2.3
pandas-datareader==0.9.0
pandasql==0.7.3
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.9.0
patsy==0.5.1
pendulum==1.4.4
pexpect==4.8.0
pgcli==1.9.0
pgspecial==1.11.5
pickleshare==0.7.5
Pillow==8.1.2
pip-autoremove==0.10.0
platformdirs==2.4.1
prison==0.2.1
prometheus-client==0.9.0
prompt-toolkit==3.0.26
property-cached==1.6.4
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-lets-be-rational==1.0.1
py-vollib==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==2.1.2
Pygments==2.8.1
PyJWT==1.7.1
pyparsing==3.0.7
pyq==5.0.0
pyrsistent==0.17.3
python-daemon==2.3.0
python-dateutil==2.8.1
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.0.3
qdldl==0.1.5.post0
qtconsole==5.2.2
QtPy==2.0.0
realpath==0.1.0
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==12.0.1
rpy2==3.4.3
rsa==4.8
scikit-learn==0.24.2
scipy==1.6.2
scs==2.1.2
seaborn==0.11.1
Send2Trash==1.8.0
setproctitle==1.2.2
shutilwhich==1.1.0
simplejson==3.17.2
six==1.16.0
sklearn==0.0
sl==1.0
sniffio==1.2.0
spotipy==2.19.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.2.4
stack-data==0.1.4
statsmodels==0.12.2
swagger-ui-bundle==0.0.9
sympy==1.8
tabulate==0.8.9
tempdir==0.7.1
tenacity==4.12.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
thrift==0.15.0
tomli==2.0.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.25.11
virtualenv==20.13.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
yfinance==0.1.59
zipp==3.7.0
zope.deprecation==4.4.0
